refactor(consolidate): extract LLM-output sanitization to consolidate/sanitize (D3 step 3)
Move recoverMalformedFrontmatter, stripOuterCodeFence, sanitizeMergedContent,
normalizeUpdatedField (+ SanitizedMergedContent) into consolidate/sanitize.ts —
pure string/frontmatter transforms with zero consolidate-domain coupling. Drops
the now-unused yaml import from the orchestrator. Re-exported; behavior-identical.
Recovery paths already pinned via consolidate-pipeline-fixes (242 tests green).
